% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{switchInput}
\alias{switchInput}
\title{<Render a Switch Input>}
\usage{
switchInput(
  inputId,
  label = c(FALSE, TRUE),
  default = FALSE,
  configuration = list()
)
}
\description{
<This input can also render any other Shiny input that is passed to it.>
}
