% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{timeInput}
\alias{timeInput}
\title{<Render a Time Input>}
\usage{
timeInput(
  inputId,
  default = format(lubridate::now(), "\%H:\%M"),
  configuration = NULL
)
}
\description{
<Based on [react-time-picker](http://projects.wojtekmaj.pl/react-time-picker/).>
}
