% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch.R
\name{switchInput}
\alias{switchInput}
\title{<Add Title>}
\usage{
switchInput(
  inputId,
  label = c(FALSE, TRUE),
  default = FALSE,
  configuration = list(),
  render_input = NULL
)
}
\description{
<Add Description>
}
